+++ src/fftln.h
@@ -2,6 +2,10 @@
 #ifndef __NTL_FFTLN_H__
 #define __NTL_FFTLN_H__
 
+#include "dft.h"
+#include "mat.h"
+#include "vec.h"
+
 /**
  * Algorithm for very large n=2^l
  */
