+++ src/gfpn.h
@@ -2,7 +2,7 @@
 #ifndef __NTL_GFPN_H__
 #define __NTL_GFPN_H__
 
-#include "ntl.h"
+#include "gf.h"
 
 template<typename T>
 class Poly;
