+++ src/misc.cpp
@@ -1,5 +1,6 @@
+#include <iostream>
 
-#include "ntl.h"
+#include "misc.h"
 
 std::ostream&
 operator<<(std::ostream& dest, __uint128_t value)
