+++ src/vecp.h
@@ -2,6 +2,12 @@
 #ifndef __NTL_VECP_H__
 #define __NTL_VECP_H__
 
+#include <algorithm>
+#include <iostream>
+#include <cassert>
+#include <cstring>
+#include <vector>
+
 /*
  * Wrapper of a vector of `n` buffers each of `size` elements of type T
  *  std::vector<T*> vec(n)
